Update setup.sh
 Allow "Lane 0" for backend

+++ setup.sh
@@ -16,10 +16,17 @@ fi
 
 # check that we're a lane
 LANEID=`hostname`
-LANENUMBER=`echo ${LANEID}|sed -n 's/^lane\([0-9]*\)$/\1/p'`
 if [ -z "$LANENUMBER" ]; then
-	echo "Host '${LANEID}' does not appear to be a POS lane. Aborting lane install." >&2
-	exit 2
+	LANENUMBER=`echo ${LANEID}|sed -n 's/^lane\([0-9]*\)$/\1/p'`
+fi
+if [ -z "$LANENUMBER" ]; then
+	read -p "Host '${LANEID}' does not appear to be a POS lane. Install backend only? [y/N] " -n 1 -r
+	echo
+	if [[ ! $REPLY =~ ^[Yy]$ ]]; then
+		echo "Aborting lane install." >&2
+		return
+	fi
+	LANENUMBER=0
 fi
 
 
@@ -60,6 +67,10 @@ if [ -n "$1" -a "$1" = "rm" ]; then
 	cd "$COREPOS"
 	git clone https://github.com/GeorgeStreetCoop/CORE-POS.git "$COREPOS"
 else
+	if [ ! -d "$COREPOS" ]; then
+		echo "Directory '$COREPOS' doesn't exist. Aborting lane install. Try again with 'rm' override parameter?" >&2
+		return
+	fi
 	cd "$COREPOS"
 	git reset --hard HEAD
 	git pull
